[bug] what's wrong with my compile

+++ include/item.h
@@ -82,7 +82,7 @@ void Item_readJSON( void *input, cJSON *jitem);
 void Item_writeJSON(void *input, cJSON *jitem);
 
 /* --- Type --- */
-int Item_Archetype(i16 id);
+u64 Item_Archetype(i16 id);
 
 /* --- Is --- */
 b32 Item_isStaff(  i16 id);
